[Features, persistence] Any atomic feature can be "underspecified" w/r/t its value, so:
- Add :ALLOWS relationship to "underspecified" value when creating a new
  atomic feature.
- Stop showing "underspecified" value in
  + global value list
  + target lists of individual features
  to make sure users can't delete (or modify) it.

+++ ode/app/controllers/Features.java
@@ -231,7 +231,7 @@ public class Features extends Controller {
             deleted.onRedeem(
                 new Callback<Boolean>() {
                     public void invoke(Boolean deleted) {
-                        if (deleted) {
+                        if (deleted && !value.name.equals("underspecified")) {
                             value.deleteIfOrphaned();
                         }
                     }
